Move sanitize_html to external.utils.functions.
