+++ components/server/src/external/routes/subject.py
@@ -9,8 +9,10 @@ from database.measurements import measurements_by_metric
 from database.reports import insert_new_report, latest_reports, metrics_of_subject
 from model.actions import copy_subject, move_item
 from model.data import ReportData, SubjectData
-from server_utilities.functions import report_date_time, sanitize_html, uuid
+from server_utilities.functions import report_date_time, uuid
 from server_utilities.type import MetricId, ReportId, SubjectId
+
+from ..utils.functions import sanitize_html
 
 from .plugins.auth_plugin import EDIT_REPORT_PERMISSION
